@@ -1,26 +1,17 @@
 module github.com/mutablelogic/go-media
 
-go 1.17
-
-replace github.com/zyedidia/highlight => github.com/djthorpe/highlight v0.0.0-20211010083339-d90b2f7f5bae
+go 1.18
 
 require (
 	github.com/djthorpe/go-errors v1.0.2
-	github.com/djthorpe/gopi/v3 v3.0.33
 	github.com/hashicorp/go-multierror v1.1.1
-	github.com/mutablelogic/go-server v1.0.37
-	github.com/olekukonko/tablewriter v0.0.4
-	golang.org/x/oauth2 v0.0.0-20211104180415-d3ed0bb246c8
+	github.com/stretchr/testify v1.8.1
+	golang.org/x/exp v0.0.0-20221212164502-fae10dda9338
 )
 
 require (
-	cloud.google.com/go v0.65.0 // indirect
-	github.com/djthorpe/go-marshaler v0.0.15 // indirect
-	github.com/golang/protobuf v1.4.3 // indirect
-	github.com/hashicorp/errwrap v1.1.0 // indirect
-	github.com/mattn/go-runewidth v0.0.9 // indirect
-	golang.org/x/net v0.0.0-20210917221730-978cfadd31cf // indirect
-	google.golang.org/appengine v1.6.6 // indirect
-	google.golang.org/protobuf v1.25.0 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/hashicorp/errwrap v1.0.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
