@@ -1,5 +1,4 @@
 /*
-Package media provides high-level media services for multiplexing, extracting and transcoding audio
-and video.
+Package media provides media open, close and conversion support
 */
 package media
